Extract point import logic into separate module
This commit includes refactor of the points module by extracting a bulk of its codebase responsible for importing points from a JSON file into a new, separate module `import.rs`. The reason behind this change is to keep the logic, focused on import-related functionality encapsulated. This results in cleaner and more maintainable code, and easier reasoning about the points module logic.