Refactor Points module and improve test coverage
The tests.rs file was deleted and its contents were moved to respective direct submodules for better organization and ease of navigation. The Points module is split now into import.rs and export.rs each responsible for importing and exporting points respectively. Functionality to export points to a JSON file was added to the export.rs module. The Cargo.toml was updated to include the serial_test crate used to serialize tests. The mod.rs file in the points directory was updated to include new struct construct and data structuring methods. This commit was necessary in order to enhance readability and maintainability of the codebase.