Add setup step for .cazan directory in workflows
This commit introduces setup steps for .cazan directory in both release and pull-request workflows. These steps create the .cazan/build directory and generate two json files inside it. This setup step was necessary to ensure the workflows have the needed directories and files for testing all features of the application.