Refactor code for better efficiency and readability
Changes were made in both import.rs and export.rs primarily for code optimization and readability. In import.rs, instead of passing a reference to the function `Self::extract_points`, the image in itself is now passed. This improves code efficiency and readability by reducing unnecessary indirections. In export.rs, a change was made from `"\n"` to `'\n'` to leverage the efficiency of character replacements instead of string in `json.replace`.